the one-arg form of getCellState is EXCLUSIVELY for changing rules
if you use it wrong, data might not behave as expected
e.g. in half-dead rules, one-arg getCellState turns a 1 into a 2,
because a 1 is usually fully alive, but in half-dead rules a two is
fully alive!  READ THE DOC AND COMMENTS

also, in changeCell, data++ was a BAD IDEA, changed to
getCellStateFromClick